ENG-3675 chore(1ui): truncate long descriptions (#831)
## Affected Packages

Apps

- [ ] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Some descriptions from the data population tool are > 266 characters,
the max description set in Portal, and they don't fit well into the
designs. This update truncates all descriptions over 266 chars in
ProfieCard. More discussion is needed as I don't think we should be
truncating long descriptions. Maybe we should have short and long
descriptions? I think it is good to have lots of metadata, a long
description is nice. But it doesn't look good in the the current design,
it isn't very readable. I almost imagine there being another card in the
main section that could have the long description. Maybe something we
should think about?

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)