ENG-4318 fix(portal): added success toast when copying share link (#856)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

When users clicked on the 'Copy Link' button in the ShareModal, there
was no visual notification or anything confirming that it had been
copied. Added a toast notification as well as an icon to the button that
changes based on a copied state. Also removed the logic that was inside
of the component and used the useCopy hook we leverage elsewhere for
consistency.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)