ENG-4066 fix(portal): create claim combobox search fix (#852)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This adds a check if the user is accessing the app from mobile and
changes the modal prop on the Popover conditionally. Setting modal to
true can cause some weird behavior on mobile. I was only able to test
this locally in the dev console, so we will want to double check that it
works once this is on staging.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)