ENG-4667 fix(portal): remove creator card from list detail (#865)
## Affected Packages

Apps

- [x] portal

Packages

- [ ] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This PR removes the creator card from the list detail page as it is
misleading. It shows the creator as the person who created the hastag
claim that is being used to generate the list, which is technically
incorrect because there is no "list" object created by any user. This
will change once we add the ability to save queries, as list are just
queries at the end of the day.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)