ENG-4506 feat(data-populator): Add eoa handling (#877)
- Adds in branching support for EOAs (external wallets) in addition to
smart wallets. If a user uses their EOA a smart wallet isn't configured
and instead we fallback to using wagmi for the interactions and client.
- We added hooks and modified the `useBatchCreateAtom` to support and
streamline this. Our playground route still exists as a specific, clear
way to test the logic.
- @0xjojikun also cleaned up and streamlined our env vars. I updated the
`Dockerfile` but didn't touch the `render.yaml` yet as I wanted to
ensure that we're ready for those changes.
- Fixes some prettier/type issues on some of the service files -- unsure
why these were suddenly failing but they're fixed now and everything
passes.
- The `pnpm` version bump wasn't intentional -- we should doublecheck
this @0xjojikun -- it seems to have been added after I reinstalled
---------

Co-authored-by: 0xjojikun <jojikun@intuition.systems>
Co-authored-by: Auroter <seangherardi@gmail.com>