ENG-4065 fix(portal): improve tag modal performance (#851)
## Affected Packages

Apps

- [x] portal

Packages

- [x] 1ui
- [ ] api
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

This PR improves tag modal performance, as well as fixes some issues
with some values related to tags. We have been relying on the embedded
tags on identities, but the data associated with those are specific to
the identities, and not the associated "has tag" claim. Rather than rely
on these embedded identities, we now make an additional API call in
route loaders, when necessary, to get the "has tag" claims associated
with the identity. This allows us to pass the claims `vault_id` directly
to the tag modal, which allows us to bypass the claim fetch. We still
have to fetch the vault details, so we are still leveraging defer and
suspense, but the loading time is much faster.

**Routes Affected:**
- /app/identity/{id}
- /app/profile/wallet/{wallet}
- /app/profile
- /readonly/profile/{wallet}
- /app/list/{id}
 
**Components:**
- List
- SaveForm
- SaveListModal
- TagsForm
- TagsModal
- TagsReview
- TagsSearchCombobox

## Screen Captures


https://www.loom.com/share/0e5fa078142e4b179e27fa9410f538b2?sid=084dc981-afef-4b16-8e3c-075728c02f87

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)