ENG-4383 feat(portal,1ui): claim & identity cards/rows (#884)
## Affected Packages

Apps

- [ ] data populator
- [x] portal
- [ ] template

Packages

- [x] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

- Adds StakeButton, StakeTVL to 1ui.
- Updates IdentityRow and ClaimRow 1ui components and removes the
versions that were in Portal.
- Updates ActivityItem component to use the new IdentityRow and
ClaimRow.
- Implements inline staking, uses data from API (claim/identity) on
initial load and then runs a fetcher to get updated data from RPC.
- Had to update various other components in 1ui to account for the
changes to some of the components related to this update.

Note: During this, I noticed that a bunch of terminology/nomenclature
used in 1ui is confusing or could be better. Once this is in I will
rebase my 1ui re-org branch and probably make a few changes to improve
this. i.e. 'numPositions' was labeled 'numFollowers'. There were a lot
of cases where we could improve some of this. I think it was just a lack
of understanding when some of these components were built out in 1ui.
Could also leverage types better as well in some places. I'm assuming
the types were added after some of these were built.

## Screen Captures


https://www.loom.com/share/8a9cb65d178848aaa12048d26d816d1b?sid=e689d1e4-1ab3-4e97-8a6b-26580b1c3fa4

https://www.loom.com/share/59978a9b08bf4a41b62ec577260efec9?sid=f2aaf1ae-3bc2-4256-9c27-1d84f94238bf

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)