ENG-4713 fix(portal,1ui): remove staking from identity rows on list detail (#914)
## Affected Packages

Apps

- [ ] data populator
- [x] portal
- [ ] template

Packages

- [x] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

** This was branched off of `vital/eng-4712-responsiveness-fixes`. Make
sure that gets merged in first, and rebase this off main before
merging**

Removes the staking buttons and context menu from IdentityRow on List
Detail route. Staking was removed because it doesn't make sense in this
context, you'd be staking on the subject identity of the has tag claim,
and we already have the save list functionality baked into this. Removed
the context button because it lives inside of IdentityRow and clashes
with the save list button. Should revisit when we get to the list and
detail page rework and create a specific row component for list details.

## Screen Captures


![image](https://github.com/user-attachments/assets/f5b5567c-af59-4248-88b0-1688cac59bc4)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)