ENG-5224 ENG-5242 feat(graphql,1ui,template): graphql package update to new schema and modified codegen (#997)
- Updates our GraphQL package to use the new schema and deployment and
brings in the fragment and query changes from the feature branches.
- Changes the codegen approach to not rely on environment variables at
all so that the build process is more robust and flexible for people
using in external apps. Uses the local schema first and then the remote
as a fallback. This will be handled in a more thorough way in an
upcoming PR.
- Updates the local `schema.graphql` to match the remote one. 
- Updates `1ui` and `template` to use fields from the updated schema.
- Note: There is an unrelated typecheck issue in Portal related to the
new Privy version that is causing the build to fail. We should handle
that in another PR that is merged in prior to this one. I don't want to
do that in this PR.